scaffolder-react: refactor for cleaner separation between stable and next
