+++ .changeset/forty-bulldogs-build.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-scaffolder-react': patch
+---
+
+The `FormProps` type is now correctly only available as an `/alpha` export.
