+++ plugins/scaffolder-react/src/index.ts
@@ -19,6 +19,6 @@ export * from './types';
 export * from './secrets';
 export * from './api';
 export * from './hooks';
+export * from './layouts';
 
 export * from './next';
-export type { FormProps } from './next';
