+++ plugins/scaffolder-react/src/next/index.ts
@@ -15,8 +15,6 @@
  */
 export * from './components';
 export * from './extensions';
-export * from './layouts';
 export * from './types';
 export * from './lib';
 export * from './hooks';
-export * from './layouts';
