+++ plugins/scaffolder-react/src/next/types.ts
@@ -13,32 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 import type { FormProps as SchemaFormProps } from '@rjsf/core-v5';
+
 /**
  * Any `@rjsf/core` form properties that are publicly exposed to the `NextScaffolderpage`
  *
- * @public
+ * @alpha
  */
 export type FormProps = Pick<
   SchemaFormProps,
   'transformErrors' | 'noHtml5Validate'
 >;
-
-/**
- * The field template from `@rjsf/core` which is a react component that gets passed `@rjsf/core` field related props.
- *
- * @public
- */
-export type LayoutTemplate<T = any> = NonNullable<
-  SchemaFormProps<T>['uiSchema']
->['ui:ObjectFieldTemplate'];
-
-/**
- * The type of layouts that is passed to the TemplateForms
- *
- * @public
- */
-export interface LayoutOptions<P = any> {
-  name: string;
-  component: LayoutTemplate<P>;
-}
